education_platform/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
psycopg2-binary==2.9.8
sqlparse==0.4.4
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0

